servicedisco: fix a bug where service registrations would be lost
This PR fixes a bug where service registrations can be lost. The implementation
of MakeAllocServiceID is faulty, and used in both Nomad and Consul service
registrations. The helper function does not take the Tags of a service into
account when generating a [nominally] unique ID for the service.

For example,

group "group" {
    service {
      name     = "db"
      tags     = ["one"]
      provider = "nomad"
    }

    service {
      name     = "db"
      tags     = ["two"]
      provider = "nomad"
    }
}

In this case only one service would remain registered - as the second one in
the list would upsert and thus overwrite the first, because they share the
same ID generated by that helper function.

The fix is to take Tags into account in the helper function. Since services are
getting a new ID anyway, take the opportunity to drive-by cleanup the output of this
helper function.